backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
numpy==2.4.4
pandas==3.0.2
matplotlib==3.10.8
scikit-learn==1.6.1
django-redis==5.4.0
redis==5.0.1
